+++ security/src/main/java/edu/cnu/swacademy/security/asset/CashWalletHistoryRepository.java
@@ -3,9 +3,9 @@ package edu.cnu.swacademy.security.asset;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CashWalletHistoryRepository extends JpaRepository<CashWalletHistory, Long> {
+public interface CashWalletHistoryRepository extends JpaRepository<CashWalletHistory, Integer> {
 
-  List<CashWalletHistory> findByCashWalletId(Long cashWalletId);
+  List<CashWalletHistory> findByCashWalletId(int cashWalletId);
 
-  List<CashWalletHistory> findByCashWalletIdOrderByCreatedAtDesc(Long cashWalletId);
+  List<CashWalletHistory> findByCashWalletIdOrderByCreatedAtDesc(int cashWalletId);
 }
