+++ security/src/main/java/edu/cnu/swacademy/security/auth/AuthenticationRepository.java
@@ -4,8 +4,8 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuthenticationRepository extends JpaRepository<Authentication, Long> {
+public interface AuthenticationRepository extends JpaRepository<Authentication, Integer> {
 
   Optional<Authentication> findByRefreshToken(String refreshToken);
-  Optional<Authentication> findByUserIdAndRefreshToken(Long userId, String refreshToken);
+  Optional<Authentication> findByUserIdAndRefreshToken(int userId, String refreshToken);
 }
